Перенесла view во ViewController
